HCS-4: add investment withdraw policy and add authorize in crud

+++ app/Http/Livewire/Investments/Withdrawals/Destroy.php
@@ -4,14 +4,19 @@ namespace App\Http\Livewire\Investments\Withdrawals;
 
 use App\Models\InvestmentWithdraw;
 use Illuminate\Contracts\View\View;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 
 class Destroy extends Component
 {
+    use AuthorizesRequests;
+
     public ?InvestmentWithdraw $investmentWithdraw = null;
 
     public function save(): void
     {
+        $this->authorize('delete', $this->investmentWithdraw);
+
         $this->investmentWithdraw->delete();
 
         $this->emit('investment::withdraw::deleted');
